Restructure integration tests to use separate JSON and corpus files
This makes these tests way easier to write and read.
