+++ spec/fixtures/test_grammars/external_extra_tokens/corpus.txt
@@ -1,0 +1,10 @@
+========================
+extra external tokens
+========================
+
+x = # a comment
+y
+
+---
+
+(assignment (variable) (comment) (variable))
