+++ spec/fixtures/test_grammars/precedence_on_single_child_positive/corpus.txt
@@ -1,0 +1,13 @@
+===========================
+function calls with blocks
+===========================
+
+foo bar { baz }
+
+---
+
+(expression (function_call
+  (identifier)
+  (expression (function_call
+    (identifier)
+    (block (expression (identifier)))))))
