+++ spec/fixtures/test_grammars/start_rule_is_blank/corpus.txt
@@ -1,0 +1,7 @@
+========================
+the empty string
+=======================
+
+---
+
+(first_rule)
