+++ spec/fixtures/test_grammars/start_rule_is_blank/grammar.json
@@ -1,0 +1,6 @@
+{
+  "name": "start_rule_is_blank",
+  "rules": {
+    "first_rule": {"type": "BLANK"}
+  }
+}
