+++ spec/fixtures/test_grammars/start_rule_is_token/corpus.txt
@@ -1,0 +1,6 @@
+===========================
+the single token
+==========================
+the-value
+---
+(first_rule)
