+++ spec/fixtures/test_grammars/start_rule_is_token/grammar.json
@@ -1,0 +1,6 @@
+{
+  "name": "start_rule_is_token",
+  "rules": {
+    "first_rule": {"type": "STRING", "value": "the-value"}
+  }
+}
