+++ spec/helpers/load_language.h
@@ -5,8 +5,10 @@
 #include "tree_sitter/runtime.h"
 #include <string>
 
-const TSLanguage *load_compile_result(const std::string &, const TSCompileResult &,
-                                      std::string external_scanner_path = "");
-const TSLanguage *get_test_language(const std::string &language_name);
+const TSLanguage *load_real_language(const std::string &name);
+
+const TSLanguage *load_test_language(const std::string &name,
+                                     const TSCompileResult &compile_result,
+                                     std::string external_scanner_path = "");
 
 #endif  // HELPERS_LOAD_LANGUAGE_H_
