+++ spec/helpers/read_test_entries.cc
@@ -1,20 +1,18 @@
 #include "helpers/read_test_entries.h"
+#include <assert.h>
 #include <string>
-#include <fstream>
-#include <streambuf>
-#include <dirent.h>
-
 #include <regex>
+#include "helpers/file_helpers.h"
+
 using std::regex;
 using std::regex_search;
 using std::regex_replace;
-using std::smatch;
 using std::regex_constants::extended;
-
+using std::smatch;
 using std::string;
 using std::vector;
-using std::ifstream;
-using std::istreambuf_iterator;
+
+string fixtures_dir = "spec/fixtures/";
 
 static string trim_output(const string &input) {
   string result(input);
@@ -27,7 +25,7 @@ static string trim_output(const string &input) {
 
 static vector<TestEntry> parse_test_entries(string content) {
   regex header_pattern("===+\n"  "([^=]+)\n"  "===+\n", extended);
-  regex separator_pattern("---+\n", extended);
+  regex separator_pattern("---+\r?\n", extended);
   vector<string> descriptions;
   vector<string> bodies;
 
@@ -55,51 +53,42 @@ static vector<TestEntry> parse_test_entries(string content) {
         body.substr(0, matches.position() - 1),
         trim_output(body.substr(matches.position() + matches[0].length()))
       });
+    } else {
+      puts(("Invalid corpus entry with description: " + descriptions[i]).c_str());
+      abort();
     }
   }
 
   return result;
 }
 
-static vector<string> list_directory(string dir_name) {
-  vector<string> result;
-
-  DIR *dir = opendir(dir_name.c_str());
-  if (!dir) {
-    printf("\nTest error - no such directory '%s'", dir_name.c_str());
-    return result;
-  }
-
-  struct dirent *dir_entry;
-  while ((dir_entry = readdir(dir))) {
-    string name(dir_entry->d_name);
-    if (name != "." && name != "..")
-      result.push_back(dir_name + "/" + name);
-  }
-
-  closedir(dir);
-  return result;
-}
-
-static string read_file(string filename) {
-  ifstream file(filename);
-  string result((istreambuf_iterator<char>(file)), istreambuf_iterator<char>());
-  return result;
-}
-
-vector<TestEntry> read_corpus_entries(string language_name) {
+vector<TestEntry> read_real_language_corpus(string language_name) {
   vector<TestEntry> result;
 
-  string fixtures_dir = "spec/fixtures/";
-
   string test_directory = fixtures_dir + "grammars/" + language_name + "/grammar_test";
-  for (string &test_filename : list_directory(test_directory))
-    for (TestEntry &entry : parse_test_entries(read_file(test_filename)))
+  for (string &test_filename : list_directory(test_directory)) {
+    for (TestEntry &entry : parse_test_entries(read_file(test_directory + "/" + test_filename))) {
       result.push_back(entry);
+    }
+  }
 
   string error_test_filename = fixtures_dir + "/error_corpus/" + language_name + "_errors.txt";
-  for (TestEntry &entry : parse_test_entries(read_file(error_test_filename)))
+  for (TestEntry &entry : parse_test_entries(read_file(error_test_filename))) {
     result.push_back(entry);
+  }
+
+  return result;
+}
+
+vector<TestEntry> read_test_language_corpus(string language_name) {
+  vector<TestEntry> result;
+
+  string test_directory = fixtures_dir + "test_grammars/" + language_name;
+  for (string &test_filename : list_directory(test_directory)) {
+    for (TestEntry &entry : parse_test_entries(read_file(test_directory + "/" + test_filename))) {
+      result.push_back(entry);
+    }
+  }
 
   return result;
 }
